Comment 2.0 You Can (Not) Read
When the threat of Tanenbaum escalates, ICS 169  is pushed to its limits, with CIDARU at the forefront of the struggle.  CIDARU are assited by two mentors from the gaming industry.  With the aid of Visual Studio, equipped with Comic Sans perfect for destroying all readers, the eight souls code desperatly to protect Gundalf and prevent an impending F.  But when startling comments are brought to light, will the heroes' greatest challenge prove to be the spaghetti code...or humanity itself?