Comment 3.0 You Can (Not) Understand
Fourteen days after Winter break, the Git is a post-apocalyptic wasteland, spaghetti code is in ruins, and the groupmates Richard knows are almost unrecognizable.  Trapped inside C#, he is recovered from Unity by anime and Smash Ultimate, only to find himself  prisoner of Comments, something that shouldn't be written.  Cold and bitter, these spaghetti code view him with suspicion and refuse to support him as he comes to terms with the consequences of reading his hard to write code.